pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[dirs.env]
virtual = ".venv"

[project]
name = "voice-stack"
version = "0.1.0"
description = "ASR + TTS services with CUDA/Metal-friendly"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Vitor Garbim", email = "vyscava@gmail.com" }]
keywords = ["speech-recognition", "whisper", "asr", "webservice"]

# Cross-platform/runtime deps only
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
  "pydantic",
  "requests",
  "ffmpeg-python"
]

[project.optional-dependencies]
# Split by feature so you can `pip install .[tts]` on lightweight machines, etc.
tts = [
  "coqui-tts",
  "resampy",
  "soundfile",
  "numpy",
  "langdetect",
]
asr = [
  "faster-whisper",
  "openai-whisper",
  "silero-vad",
  "torch",
  "torchaudio",
  "ipdb"
]
server = [
  "python-dotenv",
  "pydantic-settings",
]
dev = [
  "nox",
  "black",
  "ruff",
  "mypy",
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-mock>=3.12.0",
  "httpx>=0.26.0",
  "types-requests",
  "types-PyYAML",
  "debugpy",
  "pre-commit"
]

[tool.hatch.metadata]
allow-direct-refs = true

[tool.hatch.build.targets.sdist]
include = ["src", "pyproject.toml", "README.md", "LICENSE"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

# -----------------------
# Hatch environments
# -----------------------
[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
features = ["dev", "server", "tts", "asr"]
dependencies = []

[tool.hatch.envs.default.scripts]
fmt = [
  "black {root}/src {root}/tests {root}/scripts",
  "ruff check --fix {root}/src {root}/tests {root}/scripts",
]
lint = [
  "black --check {root}/src {root}/tests {root}/scripts",
  "ruff check {root}/src {root}/tests {root}/scripts",
]
typecheck = "mypy {root}/src {root}/scripts"

# Test commands
test = "pytest -v {args:}"
test-fast = "pytest -q --disable-warnings --maxfail=1 {args:}"
test-asr = "pytest tests/unit/asr -v {args:}"
test-tts = "pytest tests/unit/tts -v {args:}"
test-core = "pytest tests/unit/core -v {args:}"
test-utils = "pytest tests/unit/utils -v {args:}"
test-integration = "pytest tests/integration -v {args:}"

# Coverage commands
cov = "pytest --cov=src --cov-report=term-missing --cov-report=html --cov-report=xml {args:}"
cov-asr = "pytest tests/unit/asr --cov=src/asr --cov-report=term-missing {args:}"
cov-tts = "pytest tests/unit/tts --cov=src/tts --cov-report=term-missing {args:}"
cov-report = "coverage report --show-missing {args:}"
cov-html = "coverage html {args:}"

run_asr = "uvicorn asr.app:app --app-dir {root}/src --host 0.0.0.0 --port 5001 --reload --reload-dir {root}/src --log-level info"
run_tts = "uvicorn tts.app:app --app-dir {root}/src --host 0.0.0.0 --port 5002 --reload --reload-dir {root}/src --log-level info"

run_asr_dbg = "python -m debugpy --listen 0.0.0.0:5678 -m uvicorn asr.app:app --app-dir {root}/src --host 0.0.0.0 --port 5001 --log-level debug"
run_tts_dbg = "python -m debugpy --listen 0.0.0.0:5678 -m uvicorn tts.app:app --app-dir {root}/src --host 0.0.0.0 --port 5002 --log-level debug"

smoke = "python -c \"print('Add small smoke tests here')\""

# Production environments (no dev dependencies)
[tool.hatch.envs.prod-asr]
type = "virtual"
path = ".venv-asr"
features = ["server", "asr"]
dependencies = []
[tool.hatch.envs.prod-asr.scripts]
run = "uvicorn asr.app:app --app-dir {root}/src --host 0.0.0.0 --port 5001 --log-level info"

[tool.hatch.envs.prod-tts]
type = "virtual"
path = ".venv-tts"
features = ["server", "tts"]
dependencies = []
[tool.hatch.envs.prod-tts.scripts]
run = "uvicorn tts.app:app --app-dir {root}/src --host 0.0.0.0 --port 5002 --log-level info"

[tool.hatch.envs.test]
type = "virtual"
path = ".venv-test"
features = []
dependencies = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]
[tool.hatch.envs.test.scripts]
run = "pytest -q --disable-warnings --maxfail=1"
cov = "pytest --cov=src --cov-report=term-missing"

[tool.hatch.envs.typecheck]
type = "virtual"
path = ".venv-tycheck"
skip-install = true
features = []
dependencies = ["mypy>=1.10.0", "types-requests", "types-PyYAML"]
[tool.hatch.envs.typecheck.scripts]
run = "mypy src"

[tool.hatch.envs.lint]
type = "virtual"
path = ".venv-lint"
skip-install = true
features = []
dependencies = ["ruff>=0.5.0"]
[tool.hatch.envs.lint.scripts]
run =[
  "black --check {root}/src {root}tests",
  "ruff check {root}/src {root}tests",
  "mypy {root}/src"
]

[tool.hatch.envs.fmt]
type = "virtual"
path = ".venv-fmt"
skip-install = true
features = []
dependencies = ["black>=24.4.0", "ruff>=0.5.0"]
[tool.hatch.envs.fmt.scripts]
run = [
  "black {root}/src/*.py {root}/tests/*.py {root}/scripts/*.py",
  "ruff check --fix {root}/src {root}/tests {root}/scripts",
]

# -----------------------
# Ruff
# -----------------------
[tool.ruff]
target-version = "py310"
line-length = 120
extend-exclude = ["build", "dist", ".nox", ".venv", "venv"]

[tool.ruff.lint.isort]
known-first-party = ["asr", "app", "core", "tts"]
known-third-party = ["TTS"]

[tool.ruff.lint]
select = [
  "E", "F",       # pycodestyle/pyflakes
  "I",            # import sorting (used in fmt step)
  "UP",           # pyupgrade
  "B",            # bugbear
  "C4",           # comprehensions
  "PIE",          # misc correctness
  "PERF",         # performance
]
ignore = [
  # Add rule codes to ignore if needed
  "B008", # do not perform function calls in argument defaults
  "C901", # too complex
  #"UP006", # do not use typing (mypy) use annotations
]

# Use the new unified Ruff formatter only if you decide to drop Black entirely.
# [tool.ruff.format]
# docstring-code-format = true

# -----------------------
# Black
# -----------------------
[tool.black]
line-length = 120
target-version = ["py310"]
extend-exclude = "build|dist|.nox|.venv|venv"

# -----------------------
# MyPy
# -----------------------
[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
strict_equality = true
pretty = true
show_column_numbers = true
show_error_codes = true
ignore_missing_imports = true

# Per-module relaxation is handy while migrating:
[[tool.mypy.overrides]]
module = [
  "ffmpeg.*",
  "torch.*",
  "resampy.*",
  "soundfile.*",
  "numpy.*",
  "silero_vad.*"
]
ignore_missing_imports = true

# -----------------------
# Pytest
# -----------------------
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--showlocals",
]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
  "unit: Unit tests",
  "integration: Integration tests",
  "slow: Slow running tests",
  "asr: ASR service tests",
  "tts: TTS service tests",
  "requires_gpu: Tests that require GPU",
  "requires_models: Tests that require model downloads",
]
filterwarnings = [
  "error",
  "ignore::UserWarning",
  "ignore::DeprecationWarning",
]
asyncio_mode = "auto"

# -----------------------
# Coverage
# -----------------------
[tool.coverage.run]
source = ["src"]
omit = [
  "*/tests/*",
  "*/__pycache__/*",
  "*/.venv/*",
  "*/venv/*",
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
